split forms state into several uses
