+++ app/react/forms/reducers.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+
+import translations from './translations/reducer';
+
+
+export default combineReducers({
+  translations
+});
